chore: split the bft-consensus module into multiple
